feat(jsii): allow user-owned tsconfig.json files
Makes it possible for users to own more of the TypeScript configuration
by actually owing the `tsconfig.json` file instead of having it

The configuration is moved from the `jsii` key in `package.json` to the
`x-aws-jsii` key in `tsconfig.json` (which the user must then check in
with their source code, as with any other TypeScript project).

The new configuration point is entirely optional (existing projects will
continue to work the exact same way they currently do), however if there
is configuration in `tsconfig.json`, it is an error to also have
configuration in `package.json`.
