+++ packages/@scope/jsii-calc-base-of-base/.gitignore
@@ -1,4 +1,3 @@
-tsconfig.json
 dist
 .jsii
 *.tgz
