+++ packages/@scope/jsii-calc-base-of-base/package.json
@@ -35,36 +35,5 @@
     "jsii-build-tools": "^0.0.0",
     "jsii-rosetta": "^0.0.0",
     "prettier": "^2.2.1"
-  },
-  "jsii": {
-    "outdir": "dist",
-    "targets": {
-      "go": {
-        "moduleName": "github.com/aws-cdk/jsii/jsii-calc/go"
-      },
-      "java": {
-        "package": "software.amazon.jsii.tests.calculator.baseofbase",
-        "maven": {
-          "groupId": "software.amazon.jsii.tests",
-          "artifactId": "calculator-base-of-base"
-        }
-      },
-      "dotnet": {
-        "namespace": "Amazon.JSII.Tests.CalculatorNamespace.BaseOfBaseNamespace",
-        "packageId": "Amazon.JSII.Tests.CalculatorPackageId.BaseOfBasePackageId"
-      },
-      "python": {
-        "distName": "scope.jsii-calc-base-of-base",
-        "module": "scope.jsii_calc_base_of_base"
-      }
-    },
-    "versionFormat": "short",
-    "metadata": {
-      "jsii": {
-        "rosetta": {
-          "strict": true
-        }
-      }
-    }
   }
 }
