+++ packages/jsii/.gitignore
@@ -3,6 +3,7 @@
 node_modules/
 .nyc_output/
 coverage/
+schema/
 
 # copied from the root of the repo upon build
 README.md
