+++ packages/jsii/.npmignore
@@ -4,6 +4,7 @@
 !**/*.d.ts
 !bin/jsii
 !bin/jsii-fix-peers
+!schema/*.schema.json
 
 
 coverage
